kubernetes/codes/monitoring/requirements.txt
pandas
evidently
joblib
azure-storage-blob